# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aiome Management Console API",
    "description": "Core API for the Autonomous AI Operating System",
    "license": {
      "name": "ELv2",
      "identifier": "ELv2"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/skills": {
      "get": {
        "tags": [
          "crate::routes::skill"
        ],
        "operationId": "list_skills",
        "responses": {
          "200": {
            "description": "List all active skills in the Swarm",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/skills/import": {
      "post": {
        "tags": [
          "crate::routes::skill"
        ],
        "operationId": "import_skill",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Skill imported successfully"
          },
          "400": {
            "description": "Invalid URL or fetch failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Security Block (SSRF)"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ImportRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
...
      "SkillSummary": {
        "type": "object",
        "required": [
          "name",
          "description",
          "source",
          "status",
          "layer",
          "tools"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "layer": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_SERVER_SECRET"
      }
    }
  }
}